+++ bedoner/pipelines/trf_model.py
@@ -6,6 +6,8 @@ from typing import Iterable, List, Optional, Union, cast
 
 import torch
 import transformers as trf
+from transformers.modeling_xlnet import XLNET_INPUTS_DOCSTRING
+from transformers.modeling_bert import BERT_INPUTS_DOCSTRING
 from spacy.gold import GoldParse
 from spacy.language import Language
 from spacy.tokens import Doc, Span, Token
@@ -21,23 +23,40 @@ from bedoner.torch_utils import (
 )
 from bedoner.utils import zero_pad
 
-BERT_PRETRAINED_MODEL_ARCHIVE_MAP = {
-    "bert-ja-juman": "s3://bedoner/trf_models/bert/bert-ja-juman.bin"
+PRETRAINED_MODEL_ARCHIVE_MAP = {
+    "bert-ja-juman": "s3://bedoner/trf_models/bert/bert-ja-juman.bin",
+    "xlnet-ja": "s3://bedoner/trf_models/xlnet/pytorch_model.bin",
 }
-BERT_PRETRAINED_CONFIG_ARCHIVE_MAP = {
-    "bert-ja-juman": "s3://bedoner/trf_models/bert/bert-ja-juman-config.json"
+PRETRAINED_CONFIG_ARCHIVE_MAP = {
+    "bert-ja-juman": "s3://bedoner/trf_models/bert/bert-ja-juman-config.json",
+    "xlnet-ja": "s3://bedoner/trf_models/xlnet/config.json",
 }
 
 
 @dataclasses.dataclass
-class BertModelInputs:
-    """Container for BERT model input. See `trf.BertModel`'s docstring for detail."""
-
+class TransformersModelInputsBase:
     input_ids: torch.Tensor
     token_type_ids: Optional[torch.Tensor] = None
     attention_mask: Optional[torch.Tensor] = None
-    position_ids: Optional[torch.Tensor] = None
     head_mask: Optional[torch.Tensor] = None
+    __doc__ = BERT_INPUTS_DOCSTRING
+
+
+@dataclasses.dataclass
+class BertModelInputs(TransformersModelInputsBase):
+    position_ids: Optional[torch.Tensor] = None
+    __doc__ = BERT_INPUTS_DOCSTRING
+
+
+@dataclasses.dataclass
+class XLNetModelInputs(TransformersModelInputsBase):
+    mems: Optional[List[torch.FloatTensor]] = None
+    perm_mask: Optional[torch.FloatTensor] = None
+    target_mapping: Optional[torch.FloatTensor] = None
+    __doc__ = XLNET_INPUTS_DOCSTRING
+
+
+TransformersModelInputs = Union[BertModelInputs, XLNetModelInputs]
 
 
 @dataclasses.dataclass
@@ -52,15 +71,45 @@ class BertModelOutputs:
     # list of (one for each layer) of shape ``(batch_size, num_heads, sequence_length, sequence_length)``
 
 
-class BertModel(TorchPipe):
-    """Pytorch transformers BertModel component.
+@dataclasses.dataclass
+class XLNetModelOutputs:
+    """A container for trf.XLNetModel outputs. See `trf.XLNetModel`'s docstring for detail."""
 
-    Attach BERT outputs to doc.
+    laste_hidden_state: torch.FloatTensor  # shape ``(batch_size, sequence_length, hidden_size)``
+    mems: List[torch.FloatTensor]
+    hidden_states: Optional[torch.FloatTensor] = None
+    # list of (one for the output of each layer + the output of the embeddings) of shape ``(batch_size, sequence_length, hidden_size)``
+    attensions: Optional[torch.FloatTensor] = None
+    # list of (one for each layer) of shape ``(batch_size, num_heads, sequence_length, sequence_length)``
+
+
+TransformerModelOutputs = Union[BertModelOutputs, XLNetModelOutputs]
+
+
+class MODEL_NAMES:
+    bert = "bert"
+    xlnet = "xlnet"
+
+
+class CLS_NAMES:
+    inputs = "inputs"
+    outputs = "outputs"
+    model = "model"
+    config = "config"
+
+
+def get_trf_name(name: str) -> str:
+    for k in {MODEL_NAMES.bert, MODEL_NAMES.xlnet}:
+        if k in name:
+            return k
+    raise ValueError(f"Illegal model name: {name}")
+
+
+class TransformersModel(TorchPipe):
+    """Pytorch transformers Model component.
+
+    Attach the model outputs to doc.
     """
-
-    name = "bert"
-    trf_model_cls = trf.BertModel
-    trf_config_cls = trf.BertConfig
 
     def __init__(self, vocab, model=True, **cfg):
         self.vocab = vocab
@@ -80,16 +129,11 @@ class BertModel(TorchPipe):
         return cls(vocab, model=model, **cfg)
 
     @classmethod
-    def Model(cls, **cfg) -> trf.BertModel:
-        """Create `trf.BertModel`"""
+    def Model(cls, **cfg) -> trf.PreTrainedModel:
+        """Create trf Model"""
         if cfg.get("from_pretrained"):
-            cls.trf_model_cls.pretrained_model_archive_map.update(
-                BERT_PRETRAINED_MODEL_ARCHIVE_MAP
-            )
-            cls.trf_model_cls.config_class.pretrained_config_archive_map.update(
-                BERT_PRETRAINED_CONFIG_ARCHIVE_MAP
-            )
-            model = cls.trf_model_cls.from_pretrained(cfg.get("trf_name"))
+            trf_name = cfg.get("trf_name", "")
+            model = cls.trf_model_cls.from_pretrained(trf_name)
         else:
             if "vocab_size" in cfg["trf_config"]:
                 vocab_size = cfg["trf_config"]["vocab_size"]
@@ -101,23 +145,29 @@ class BertModel(TorchPipe):
     def max_length(self) -> int:
         return self.model.config.max_position_embeddings
 
-    def assert_length(self, x: BertModelInputs):
-        if x.input_ids.shape[1] > self.max_length:
+    def assert_length(self, x: TransformersModelInputs):
+        if self.max_length > 0 and x.input_ids.shape[1] > self.max_length:
             raise ValueError(
                 f"Too long input_ids. Expected {self.max_length}, but got {x.input_ids.shape[1]}"
             )
 
-    def predict(self, docs: List[Doc]) -> BertModelOutputs:
+    def predict(self, docs: List[Doc]) -> TransformerModelOutputs:
         self.require_model()
         self.model.eval()
         x = self.docs_to_trfinput(docs)
         self.assert_length(x)
         with torch.no_grad():
-            y = BertModelOutputs(*self.model(**dataclasses.asdict(x)))
+            y = self.output_cls(*self.model(**dataclasses.asdict(x)))
         return y
 
-    def set_annotations(self, docs: List[Doc], outputs: BertModelOutputs) -> None:
-        """Assign the extracted features to the Doc."""
+    def set_annotations(
+        self, docs: List[Doc], outputs: TransformerModelOutputs, set_vector: bool = True
+    ) -> None:
+        """Assign the extracted features to the Doc.
+
+        Args:
+            set_vector: If True, attach vector to doc. This may harms speed.
+        """
         for i, doc in enumerate(docs):
             length = len(doc._.trf_word_pieces)
             # Instead of assigning tensor directory, assign `TensorWrapper`
@@ -125,22 +175,6 @@ class BertModel(TorchPipe):
             doc._.trf_last_hidden_state = TensorWrapper(
                 outputs.laste_hidden_state, i, length
             )
-            doc._.trf_pooler_output = TensorWrapper(outputs.pooler_output, i, length)
-            lh: torch.Tensor = doc._.get(ATTRS.last_hidden_state).get()
-
-            doc_tensor = lh.new_zeros((len(doc), lh.shape[-1]))
-            # TODO: Inefficient
-            # TODO: Store the functionality into user_hooks after https://github.com/explosion/spaCy/issues/4439 was released
-            for i, a in enumerate(doc._.get(ATTRS.alignment)):
-                doc_tensor[i] += lh[a].sum(0)
-            doc.tensor = doc_tensor
-            doc.user_hooks["vector"] = get_doc_vector_via_tensor
-            doc.user_span_hooks["vector"] = get_span_vector_via_tensor
-            doc.user_token_hooks["vector"] = get_token_vector_via_tensor
-            doc.user_hooks["similarity"] = get_similarity
-            doc.user_span_hooks["similarity"] = get_similarity
-            doc.user_token_hooks["similarity"] = get_similarity
-
             if outputs.hidden_states:
                 doc._.trf_all_hidden_states = [
                     TensorWrapper(hid_layer, i)
@@ -152,18 +186,35 @@ class BertModel(TorchPipe):
                     for attention_layer in cast(Iterable, outputs.attensions)
                 ]
 
+            if set_vector:
+                lh: torch.Tensor = doc._.get(ATTRS.last_hidden_state).get()
+                doc_tensor = lh.new_zeros((len(doc), lh.shape[-1]))
+                # TODO: Inefficient
+                # TODO: Store the functionality into user_hooks after https://github.com/explosion/spaCy/issues/4439 was released
+                for i, a in enumerate(doc._.get(ATTRS.alignment)):
+                    doc_tensor[i] += lh[a].sum(0)
+                doc.tensor = doc_tensor
+                doc.user_hooks["vector"] = get_doc_vector_via_tensor
+                doc.user_span_hooks["vector"] = get_span_vector_via_tensor
+                doc.user_token_hooks["vector"] = get_token_vector_via_tensor
+                doc.user_hooks["similarity"] = get_similarity
+                doc.user_span_hooks["similarity"] = get_similarity
+                doc.user_token_hooks["similarity"] = get_similarity
+
     def update(self, docs: List[Doc], golds: List[GoldParse]):
         """Simply forward docs in training mode."""
         self.require_model()
         self.model.train()
         x = self.docs_to_trfinput(docs)
         self.assert_length(x)
-        y = BertModelOutputs(*self.model(**dataclasses.asdict(x)))
-        self.set_annotations(docs, y)
+        y = self.output_cls(*self.model(**dataclasses.asdict(x)))
+        # set_vector=False because vector may be not need in updating.
+        # You can still use model outputs via doc._.trf_last_hidden_state etc.
+        self.set_annotations(docs, y, set_vector=False)
 
-    def docs_to_trfinput(self, docs: List[Doc]) -> BertModelInputs:
+    def docs_to_trfinput(self, docs: List[Doc]) -> TransformersModelInputs:
         """Generate input data for trf model from docs."""
-        inputs = BertModelInputs(
+        inputs = self.input_cls(
             input_ids=torch.tensor(
                 zero_pad([doc._.trf_word_pieces for doc in docs]), device=self.device
             )
@@ -198,13 +249,37 @@ class BertModel(TorchPipe):
         with (path / "vocab.pkl").open("wb") as f:
             pickle.dump(self.vocab, f)
 
-    def from_disk(self, path: Path, exclude=tuple(), **kwargs) -> "BertModel":
+    def from_disk(self, path: Path, exclude=tuple(), **kwargs) -> "TransformersModel":
         with (path / "cfg.pkl").open("rb") as f:
             self.cfg = pickle.load(f)
         with (path / "vocab.pkl").open("rb") as f:
             self.vocab = pickle.load(f)
         self.model = self.trf_model_cls.from_pretrained(path)
         return self
+
+
+class BertModel(TransformersModel):
+    name = "bert"
+    output_cls = BertModelOutputs
+    input_cls = BertModelInputs
+    trf_model_cls = trf.BertModel
+    trf_config_cls = trf.BertConfig
+
+    def set_annotations(
+        self, docs: List[Doc], outputs: BertModelOutputs, set_vector=True
+    ) -> None:
+        super().set_annotations(docs, outputs, set_vector=set_vector)
+        for i, doc in enumerate(docs):
+            length = len(doc._.trf_word_pieces)
+            doc._.trf_pooler_output = TensorWrapper(outputs.pooler_output, i, length)
+
+
+class XLNetModel(TransformersModel):
+    name = "xlnet"
+    trf_model_cls = trf.XLNetModel
+    trf_config_cls = trf.XLNetConfig
+    input_cls = XLNetModelInputs
+    output_cls = XLNetModelOutputs
 
 
 def get_doc_vector_via_tensor(doc) -> torch.Tensor:
@@ -226,3 +301,4 @@ def get_similarity(o1: Union[Doc, Span, Token], o2: Union[Doc, Span, Token]) -> 
 
 
 Language.factories[BertModel.name] = BertModel
+Language.factories[XLNetModel.name] = XLNetModel