+++ tests/conftest.py
@@ -1,10 +1,16 @@
+from bedoner.lang.trf_mixin import TransformersLanguageMixin
+from bedoner.utils import inject_mixin
 from pathlib import Path
+import sentencepiece as spm
 
 import pytest
+from spacy.vocab import Vocab
 
 from bedoner.lang.juman import Japanese as Juman
 from bedoner.lang.knp import Japanese as KNP
 from bedoner.lang.mecab import Japanese as Mecab
+from bedoner.pipelines.wordpiecer import WordPiecer
+from bedoner.pipelines.trf_model import XLNetModel
 
 from .utils import check_juman, check_knp, check_mecab
 
@@ -39,9 +45,54 @@ def DATADIR():
 
 @pytest.fixture(scope="session")
 def fixture_dir():
-    return Path(__file__).parent / "fixtures/"
+    return (Path(__file__).parent / "fixtures/").absolute()
 
 
 @pytest.fixture(scope="session")
 def bert_dir(fixture_dir):
     return str(fixture_dir / "bert")
+
+
+@pytest.fixture(scope="session")
+def xlnet_dir(fixture_dir):
+    return str(fixture_dir / "xlnet")
+
+
+@pytest.fixture(scope="session")
+def xlnet_wp(xlnet_dir):
+    return WordPiecer.from_pretrained(Vocab(), xlnet_dir)
+
+
+@pytest.fixture(scope="session")
+def xlnet_model(xlnet_dir):
+    return XLNetModel.from_pretrained(Vocab(), xlnet_dir)
+
+
+@pytest.fixture(scope="session")
+def trf_mecab():
+    return inject_mixin(TransformersLanguageMixin, Mecab)
+
+
+@pytest.fixture(scope="session", params=["bert", "xlnet"])
+def trf_name(request):
+    return request.param
+
+
+@pytest.fixture(scope="session")
+def trf_dir(trf_name, bert_dir, xlnet_dir):
+    if trf_name == "bert":
+        return bert_dir
+    if trf_name == "xlnet":
+        return xlnet_dir
+
+
+@pytest.fixture(scope="session")
+def spiece_path(fixture_dir):
+    return str(fixture_dir / "spiece.model")
+
+
+@pytest.fixture(scope="session")
+def spiece(spiece_path):
+    s = spm.SentencePieceProcessor()
+    s.load(spiece_path)
+    return s