+++ tests/pipelines/test_trf_ner.py
@@ -7,37 +7,41 @@ import torch
 from spacy.gold import GoldParse
 from spacy.language import Language
 
-from bedoner.models import bert_ner
+from bedoner.models import trf_ner
 from bedoner.ner_labels.labels_ene import ALL_LABELS as enes
 from bedoner.ner_labels.labels_irex import ALL_LABELS as irexes
 from bedoner.ner_labels.utils import make_biluo_labels
 
 
-@pytest.fixture(scope="module")
-def labels_ene():
-    return make_biluo_labels(enes)
+label_types = ["ene", "irex"]
+
+
+@pytest.fixture(scope="module", params=label_types)
+def label_type(request):
+    return request.param
 
 
 @pytest.fixture(scope="module")
-def labels_irex():
-    return make_biluo_labels(irexes)
+def labels(label_type):
+    if label_type == "ene":
+        return make_biluo_labels(enes)
+    elif label_type == "irex":
+        return make_biluo_labels(irexes)
+    else:
+        raise ValueError
 
 
-@pytest.fixture(scope="module", params=["mecab", "juman"], ids=["mecab", "juman"])
-def nlp(labels_ene, request, bert_dir):
+@pytest.fixture(scope="module", params=["mecab", "juman", "sentencepiece"])
+def nlp(labels, request, trf_dir):
     lang = request.param
-    _nlp = bert_ner(lang=lang, labels=["-"] + labels_ene, pretrained=bert_dir)
+    # if lang == "sentencepiece" and "xlnet" not in trf_dir:
+    #     pytest.skip()
+    _nlp = trf_ner(lang=lang, labels=["-"] + labels, pretrained=trf_dir)
     assert _nlp.meta["lang"] == lang
     return _nlp
 
 
-@pytest.fixture(scope="module", params=["mecab", "juman"], ids=["mecab", "juman"])
-def nlp_irex(labels_irex, request, bert_dir):
-    lang = request.param
-    return bert_ner(lang=lang, labels=["-"] + labels_irex, pretrained=bert_dir)
-
-
-TESTCASE = [
+TESTCASE_ENE = [
     (
         "ＥＸＩＬＥのＡＴＳＵＳＨＩと中島美嘉が１４日ニューヨーク入り",
         {
@@ -60,8 +64,10 @@ TESTCASE = [
 ]
 
 
-@pytest.mark.parametrize("text,gold", TESTCASE)
-def test_call(nlp: Language, text, gold):
+@pytest.mark.parametrize("text,gold", TESTCASE_ENE)
+def test_call(nlp: Language, text, gold, label_type):
+    if label_type == "irex":
+        pytest.skip()
     nlp(text)
 
 
@@ -69,8 +75,10 @@ def test_pipe(nlp: Language):
     list(nlp.pipe(["今日はいい天気なので外で遊びたい", "明日は晴れ"]))
 
 
-@pytest.mark.parametrize("text,gold", TESTCASE)
-def test_update(nlp: Language, text, gold):
+@pytest.mark.parametrize("text,gold", TESTCASE_ENE)
+def test_update(nlp: Language, text, gold, label_type):
+    if label_type == "irex":
+        pytest.skip()
     assert nlp.device.type == "cpu"
     doc = nlp(text)
     gold = GoldParse(doc, **gold)
@@ -83,33 +91,30 @@ def test_update(nlp: Language, text, gold):
     assert doc._.loss
 
 
-def test_update_batch(nlp: Language):
-    texts, golds = zip(*TESTCASE)
+def test_update_batch(nlp: Language, label_type):
+    if label_type == "irex":
+        pytest.skip()
+    texts, golds = zip(*TESTCASE_ENE)
     optim = nlp.resume_training()
     nlp.update(texts, golds, optim)
 
 
-def test_evaluate(nlp: Language):
-    nlp.evaluate(TESTCASE)
+def test_evaluate(nlp: Language, label_type):
+    if label_type == "irex":
+        pytest.skip()
+    nlp.evaluate(TESTCASE_ENE)
 
 
-def test_save_and_load(nlp: Language):
+def test_save_and_load(nlp: Language, label_type):
+    if label_type == "irex":
+        pytest.skip()
     with tempfile.TemporaryDirectory() as d:
         nlp.to_disk(d)
         nlp = spacy.load(d)
-        nlp(TESTCASE[0][0])
+        nlp(TESTCASE_ENE[0][0])
 
 
 TESTCASE2 = ["資生堂の香水-禅とオードパルファンＺＥＮの違いを教えて下さい。また今でも製造されてますか？"]
-
-
-@pytest.mark.parametrize("text", TESTCASE2)
-def test_irex_call(nlp_irex: Language, text):
-    nlp_irex(text)
-
-
-def test_pipe_irex(nlp_irex: Language):
-    list(nlp_irex.pipe(["今日はいい天気なので外で遊びたい", "明日は晴れ"]))
 
 
 @pytest.fixture
@@ -118,15 +123,19 @@ def cuda():
 
 
 @pytest.mark.skipif(not torch.cuda.is_available(), reason="cuda test")
-@pytest.mark.parametrize("text,gold", TESTCASE)
-def test_call_cuda(nlp: Language, text, gold, cuda):
+@pytest.mark.parametrize("text,gold", TESTCASE_ENE)
+def test_call_cuda(nlp: Language, text, gold, cuda, label_type):
+    if label_type == "irex":
+        pytest.skip()
     nlp.to(cuda)
     nlp(text)
 
 
 @pytest.mark.skipif(not torch.cuda.is_available(), reason="cuda test")
-@pytest.mark.parametrize("text,gold", TESTCASE)
-def test_update_cuda(nlp: Language, text, gold, cuda):
+@pytest.mark.parametrize("text,gold", TESTCASE_ENE)
+def test_update_cuda(nlp: Language, text, gold, cuda, label_type):
+    if label_type == "irex":
+        pytest.skip()
     nlp.to(cuda)
     doc = nlp(text)
     gold = GoldParse(doc, **gold)
@@ -139,65 +148,45 @@ def test_update_cuda(nlp: Language, text, gold, cuda):
 
 
 @pytest.mark.skipif(not torch.cuda.is_available(), reason="cuda test")
-def test_update_batch_cuda(nlp: Language, cuda):
+def test_update_batch_cuda(nlp: Language, cuda, label_type):
+    if label_type == "irex":
+        pytest.skip()
     nlp.to(cuda)
-    texts, golds = zip(*TESTCASE)
+    texts, golds = zip(*TESTCASE_ENE)
     optim = nlp.resume_training()
     nlp.update(texts, golds, optim)
 
 
-@pytest.fixture(scope="module", params=["ner/ner.json"])
-def example_irex(request, DATADIR):
-    with (DATADIR / request.param).open() as f:
-        d = json.load(f)
-    return d
+@pytest.fixture(
+    scope="module",
+    params=["ner/ner-ene.json", "ner/ner-irex.json", "ner/ner-ene2.json"],
+)
+def example_gold(request, DATADIR, label_type):
+    fname = request.param
+    if label_type in fname:
+        with (DATADIR / fname).open() as f:
+            d = json.load(f)
+        return d
+    else:
+        pytest.skip()
 
 
-@pytest.fixture(scope="module", params=["ner/ner2.json"])
-def example_ene(request, DATADIR):
-    with (DATADIR / request.param).open() as f:
-        d = json.load(f)
-    return d
+@pytest.fixture(scope="module", params=["ner/ner-irex-long.json"])
+def example_long(request, DATADIR, label_type, trf_name):
+    fname = request.param
+    if trf_name == "bert" and label_type in fname:
+        with (DATADIR / fname).open() as f:
+            d = json.load(f)
+        return d
+    else:
+        pytest.skip()
 
 
-@pytest.fixture(scope="module", params=["ner/ner_ene.json"])
-def example_ene2(request, DATADIR):
-    with (DATADIR / request.param).open() as f:
-        d = json.load(f)
-    return d
-
-
-@pytest.fixture(scope="module", params=["ner/long.json"])
-def example_long(request, DATADIR):
-    with (DATADIR / request.param).open() as f:
-        d = json.load(f)
-    return d
-
-
-def test_example_batch_irex(nlp_irex: Language, example_irex):
-    texts, golds = zip(*example_irex)
-    optim = nlp_irex.resume_training()
-    nlp_irex.update(texts, golds, optim)
-
-
-def test_example_batch_ene(nlp: Language, example_ene):
-    texts, golds = zip(*example_ene)
+def test_example_batch(nlp: Language, example_gold):
+    texts, golds = zip(*example_gold)
     optim = nlp.resume_training()
     nlp.update(texts, golds, optim)
 
 
-def test_long_input(nlp: Language, example_long):
-    texts, golds = zip(*example_long)
-    optim = nlp.resume_training()
-    with pytest.raises(ValueError):
-        nlp.update(texts, golds, optim)
-
-
-def test_example_batch_ene2(nlp: Language, example_ene2):
-    texts, golds = zip(*example_ene2)
-    optim = nlp.resume_training()
-    nlp.update(texts, golds, optim)
-
-
-def test_example_batch_ene2_eval(nlp: Language, example_ene2):
-    nlp.evaluate(example_ene2)
+def test_example_batch_eval(nlp: Language, example_gold):
+    nlp.evaluate(example_gold)